HAPA-BACKEND/apply_migration.py
from extensions import supabase_client
from supabase import create_client
import os
from dotenv import load_dotenv

load_dotenv(".env")

url = os.environ.get("SUPABASE_URL")
key = os.environ.get("SUPABASE_SERVICE_KEY")

if not url or not key:
    print("Error: Missing SUPABASE_URL or SUPABASE_SERVICE_KEY in backend/.env")
    exit(1)

supabase = create_client(url, key)

with open("migrations/04_allow_null_phone.sql", "r") as f:
    sql = f.read()

print(f"Executing SQL: {sql}")

# Use the 'postgres_meta' or direct SQL execution if possible. 
# Since we might not have direct SQL access via client, we might need to use a different approach or ask user.
# However, standard Supabase client doesn't expose raw SQL execution easily unless via RPC or specific setup.
# Let's try to use the REST API 'rpc' if there is a 'exec_sql' function, OR better, 
# since we are in a dev environment and have the service key, we CANNOT run arbitrary SQL via the JS/Python client usually 
# unless there is a specific stored procedure for it.

# WAIT. Standard Supabase projects usually don't allow raw SQL from client.
# But wait, if we have the SERVICE_KEY, we might?
# Actually, the python client is just a wrapper around PostgREST. PostgREST doesn't support raw SQL.
# 
# ALTERNATIVE:
# We can tell the user to run it.
# OR
# We can try to use `psycopg2` if available (it might be in requirements).

print("Can't execute raw SQL via standard Supabase Client without an RPC.")
print("Please run the following SQL in your Supabase SQL Editor:")
print(sql)
